refactor(bottle): makeQueryClient on the server
