+++ apps/bottle/src/store/query/QueryClientProvider.tsx
@@ -1,29 +1,8 @@
 'use client';
 
-import { isServer, QueryClient, QueryClientProvider as _QueryClientProvider } from '@tanstack/react-query';
+import { QueryClientProvider as _QueryClientProvider } from '@tanstack/react-query';
 import { ReactNode } from 'react';
-
-function makeQueryClient() {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: Infinity,
-        gcTime: Infinity,
-      },
-    },
-  });
-}
-
-let browserQueryClient: QueryClient | undefined = undefined;
-
-function getQueryClient() {
-  if (isServer) {
-    return makeQueryClient();
-  } else {
-    if (!browserQueryClient) browserQueryClient = makeQueryClient();
-    return browserQueryClient;
-  }
-}
+import { getQueryClient } from './getQueryClient';
 
 export function QueryClientProvider({ children }: { children: ReactNode }) {
   const queryClient = getQueryClient();
