+++ apps/bottle/src/store/query/getQueryClient.ts
@@ -1,0 +1,32 @@
+import {
+  isServer,
+  QueryClient,
+  QueryClientProvider as _QueryClientProvider,
+  defaultShouldDehydrateQuery,
+} from '@tanstack/react-query';
+
+const DEFAULT_STALE_TIME = 10 * 60 * 1000;
+
+function makeQueryClient() {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: DEFAULT_STALE_TIME,
+      },
+      dehydrate: {
+        shouldDehydrateQuery: query => defaultShouldDehydrateQuery(query) || query.state.status === 'pending',
+      },
+    },
+  });
+}
+
+let browserQueryClient: QueryClient | undefined = undefined;
+
+export function getQueryClient() {
+  if (isServer) {
+    return makeQueryClient();
+  } else {
+    if (!browserQueryClient) browserQueryClient = makeQueryClient();
+    return browserQueryClient;
+  }
+}
